refactor: Update product.service.ts to include updateImageProduct function
The code changes add the updateImageProduct function to the product.service.ts file. This function allows updating the image of a product by making a POST request to the "/products/update-image/{productId}" endpoint with the updated image data. This enhancement improves the functionality of the product service by providing the ability to update product images.